swh.deposit.api: Add service to clean up temporary archives
The retrieval of archives can temporarily build an archive from
multiple ones before sending its content to the client.  As a result,
when retrieving a content, we leave in a temporary location an
archive.  This temporary archive is referenced in a table
deposit_temporary_archive.

This adds a service which is in charge to poll regularly and delete
old temporary archives.