swh.deposit.api: Add post check validation on deposit
Separated from the query/response flow, we now check that we can at
least read deposit's archives (implementation wise, we use django's
signal api).

If the archives' check fails, the deposit is deemed 'rejected'.  The
client can check the state endpoint to know the status.
deposit_status key).  Otherwise, the status is set to
ready-for-injection.