Move load-deposit task creation code in the deposit-check private API endpoint
Also replace utils.origin_url_from() by a Deposit.origin_url property, and
ensure the archive file format of the uploaded file is supported.

The rename of invalid.tar.gz as invalid.gz is required for tests to pass
which should not be the case. It will be investigated in a later revision,
but for now we want tests to be green.

This requires to make config dict available in SWHPrivateAPIView so we can
have access to a scheduler from there.