Pass collection + id to the checker instead of an URL.
This allows the checker to use a configurable URL, in case it
needs an URL different from the one used by the client.

For example a client outside docker would use http://localhost:5080/deposit/
to access the API, but the checker cannot use this URL (it needs either
nginx or swh-deposit as hostname).