Migrate deposit SWHIDs (data) to the new specification
Migrate both "recent" and "old" format deposits [1] to the new specification.

That means the deposit swh_id* fields will be set to:
- swh_id: directory SWHID (no context)
- swh_id_context: directory SWHID (with context, origin, visit, anchor path)

Optionally, those 2 fields will be kept (for now) and realigned where it was not
set ("old" deposits) to:
- swh_anchor_id: revision SWHID (no context)
- swh_anchor_id_context: revision SWHID (context with only origin)

It's expected some very "old" deposits won't be migrated as we cannot resolve
those values. They will be rescheduled when it will be possible to do
so (deploy [2]).

[1] "recent" format means all swh_id fields are set:
- swh_id: directory SWHID (no context)
- swh_id_context: directory SWHID (context with only origin)
- swh_anchor_id: revision SWHID (no context)
- swh_anchor_id_context: revision SWHID (context with only origin)

"old" format:
- swh_id: revision SWHID (no context)
- swh_id_context: not set
- swh_anchor_id: not set
- swh_anchor_id_context: not set

[2] Related to D3141

Related to T2398