Transit raw metadata to the loader to unify with metadata update scenario
The new update scenario now stores new metadata update to the metadata storage.
The loader does not, it currently stores the transformed xml (into json dict).
The loader passes by this deposit_read call to actually retrieve the data.

So prior to adapting the loader, the information returned by deposit_read need
to provide the raw metadata as well.

This also:
- adds type to impacted methods along the way.
- simplifies a bit the current deposit_read tests

Related to T2649