Explicitly use the atom: prefix internally.
This mostly does not change the protocol used (except in the error messages),
it's just an internal change for the server and for the client.

The only change in the protocol is that local tags (eg. `<entry>...</entry>`)
are no longer assumed to be in the Atom namespace (like
`<entry xmlns="http://www.w3.org/2005/Atom">...</entry>`), but they
never should have been in the first place.

Default namespaces / unprefixed tags are a footgun because it's too
easy to add tags in the default namespace without noticing,
or use the wrong namespace.