conftest: Set HTTP_IN_PROGRESS to True when creating partial deposits.
conftest is cheating by hard-resetting the deposit status to partial
after creating a complete deposit; but it doesn't make sense for
a real scenario.

Additionally, a future commit will add a test relying on the
'partial_deposit' not scheduling a task, so the HTTP_IN_PROGRESS
header will need to be True at that point.