Accept <codemeta:name> and <codemeta:author> as alternatives to <atom:name>/<atom:title> and <atom:author>.
This was broken by a8e86a92eeeba4ed5114ff2ef4096ba25213d98a,
as the check_metadata() checks whether there is a tag that *contains*
the expected name, checking for 'author' (the old name of 'atom:author')
accidentally matched 'codemeta:author' as well.

This resulted in the right behavior in the majority of the cases
(accepted 'codemeta:author'), but for the wrong reason, and explicitly
renaming to atom:author broke this.

Ditto for name.

A future commit will remove the substring matching to remove
false positives (eg. 'atom:authorblahblah' should not be accepted
as 'atom:author')