Allow to configure authentication mechanism per config file
This will allow to deploy on staging with keycloak while continue using the deposit with
http basic authentication in production.

If some urgency fix is needed, we won't be forced into migrating keycloak immediately to
fix that potential other issue.

Note that the following transitional code will probably go away when we are done
migrating to keycloak.

Related to T2858