Open a paginated list user deposits endpoint
This adds a paginated listing endpoint so authenticated user can retrieve their deposit
information in batch. This touches another part of an equivalent private listing api to
allow paginated code reuse.

That new endpoint is not a sword endpoint but it lists deposits in xml relatively sword
like nonetheless.

Related to T2996