Open `swh deposit list` cli to allow user to list their deposits
This does not provide any filters yet though. The result is paginated and "next" link
will be displayed in the result. It's up to clients to make the subsequent calls to the
cli to list the next deposits through the --page and --page-size flags.

```
$ swh deposit list --help
Usage: swh deposit list [OPTIONS]

  Deposit metadata only upload

Options:
  --url TEXT                      (Optional) Deposit server api endpoint. By
                                  default,
                                  https://deposit.softwareheritage.org/1

  --username TEXT                 (Mandatory) User's name  [required]
  --password TEXT                 (Mandatory) User's associated password
                                  [required]

  -f, --format [logging|yaml|json]
                                  Output format results.
  --page INTEGER                  Page number when requesting more information
  --page-size INTEGER             Page number when requesting more information
  -h, --help                      Show this message and exit.
```

This allows to list from the cli the user's deposits (no filter).

Related to T2996