deposit_check: Actually store warning in deposit status detail
Prior to this commit, only rejected deposit were storing problem details. Now that we
can have warnings even in case of 'verified' deposit, we need to store that details for
post-analysis.

Note that this also fixes the docstring of the overall class which were out of date
since the beginning (duplicated from another class).

Related to T3677