deposit_list: Only enrich deposit metadata in page
Prior to this, this was fetching all deposits and then for each deposit, query further
information. Then return results and let the pagination happen.

This now keeps the queryset lazy, the pagination happens and when a page is requested,
this fetches further information on the subset required.

Related to T4020