model: replace contrib's JSONField by a custom one
That uses django.db.models' by default, and fall back to the old contrib
one if need be.

This later has been deprecated for some time and is now producing
errors.

Doing this should work in any situation. The migration is actually only
a matter of the name of the type in the field metadata, so using a
stable one (swh.deposit.model.JSONField) make it backend-agnostic.