api/private: Add endpoint to get download links of uploaded tarballs
Add a new private Web API endpoint to get a list of URLs for downloading
the tarballs uploaded with a deposit.

In development or docker mode, the tarballs are stored in the local
filesystem and served by django.

In production mode, the tarballs are stored in an azure blob storage
and temporary download links with a shared access signature are
generated when requesting the endpoint.

It enables to move costly operations related to downloading and processing
tarballs in celery workers instead of letting the deposit server performing
those tasks.

Related to #4657 and #4658.