checker: Remove private API endpoint and do checks on celery worker
Checking deposit archives can be a costly operation as the checker
must download the archives to list their content.

It has been observed in production that if a large archive has been
uploaded with a deposit, requesting the check endpoint of the private
deposit API can end up with gunicorn worker being killed as the
time to download the archive exceeds the worker timeout.

So instead of using the private API endpoint performs the checks,
prefer to move these operations in the celery worker executing the
check-deposit task.

Fixes #4658.