tests/cli/test_client: Fix failing tests with click 8.3.0
A race condition in click.testing.StreamMixer was fixed in click 8.2.2
that makes some tests fail due to in/out streams used by the CliRunner
being closed early. It has a side effect of making logging fails
afterwards as stderr stream is already closed.

Fix the failures by checking content of the stdout stream instead of
the output one that contains the content of both stdout and stderr.

We are only hit by that bug now as we previously forbid the use of
click 8.2.2 due to another bug.